[Footer] refactoring: extract animation header wrapper to other client component
