+++ components/Header/Header.component.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import s from "./Header.component.module.scss";
 import HamburgerLogo from "./HamburgerLogo/HamburgerLogo.component";
 
@@ -7,34 +5,12 @@ import logo from "@/assets/images/logo.jpeg";
 
 import SearchSetion from "./SearchSection/SearchSetion";
 import Image from "next/image";
-import { useContext, useEffect, useRef, useState } from "react";
-import HamburgerClickContext from "@/store/HamburgerClickContext";
+import HeaderAnimationWrapper from "./HeaderAnimationWrapper/HeaderAnimationWrapper.component";
 
 const Header = () => {
-  const { isOpen } = useContext(HamburgerClickContext);
-
-  const header = useRef<HTMLDivElement>(null);
-  const [isSticky, setIsSticky] = useState(false);
-  const handleScroll = () => {
-    if (header.current) {
-      if (scrollY > 46 && !isOpen) {
-          setIsSticky(true);
-      } else {
-        setIsSticky(false);
-      }
-    }
-  };
-
-
-  useEffect(() => {
-    addEventListener("scroll", handleScroll);
-    return () => {
-      removeEventListener("scroll", handleScroll);
-    };
-  }, []);
   return (
     <>
-      <header ref={header} className={` ${isSticky ? s.sticky : s.header}`}>
+      <HeaderAnimationWrapper stickyClass={s.sticky} normalClass={s.header}>
         <section className={s.header__logoSection}>
           <Image src={logo} alt="logo of company" width={100} height={100} />
         </section>
@@ -53,9 +29,7 @@ const Header = () => {
           </div>
         </section>
         <HamburgerLogo />
-
-        {/* Hamburger Menu */}
-      </header>
+      </HeaderAnimationWrapper>
     </>
   );
 };
