+++ components/Header/HeaderAnimationWrapper/HeaderAnimationWrapper.component.tsx
@@ -1,0 +1,44 @@
+"use client";
+
+import HamburgerClickContext from "@/store/HamburgerClickContext";
+import { type ReactNode, useContext, useEffect, useRef, useState } from "react";
+
+type HeaderAnimationWrapperProps = {
+  children: ReactNode;
+  stickyClass: string;
+  normalClass: string;
+};
+
+function HeaderAnimationWrapper({
+  children,
+  stickyClass,
+  normalClass,
+}: HeaderAnimationWrapperProps) {
+  const { isOpen } = useContext(HamburgerClickContext);
+
+  const header = useRef<HTMLDivElement>(null);
+  const [isSticky, setIsSticky] = useState(false);
+  const handleScroll = () => {
+    if (header.current) {
+      if (scrollY > 46 && !isOpen) {
+        setIsSticky(true);
+      } else {
+        setIsSticky(false);
+      }
+    }
+  };
+
+  useEffect(() => {
+    addEventListener("scroll", handleScroll);
+    return () => {
+      removeEventListener("scroll", handleScroll);
+    };
+  }, []);
+  return (
+    <header ref={header} className={` ${isSticky ? stickyClass : normalClass}`}>
+      {children}
+    </header>
+  );
+}
+
+export default HeaderAnimationWrapper;
